Update downloader.py
If the user chooses multiple songs of one album this gets recognised and in the album menu they are now all checked as expected.
One small fix made, that if genius were to name the artist differently in the artist list than in the prymary artist tag it will not attempt to rearange the list.